validation

+++ server/routes/rogals.js
@@ -261,10 +261,10 @@ router.delete('/:id', auth, async (req, res) => {
             return res.status(404).json({msg: 'Rogal not found'});
         }
 
-        // // Check user
-        // if (rogal.user.toString() !== req.user.id) {
-        //     return res.status(401).json({msg: 'User not authorized'});
-        // }
+        // Check user
+        if (rogal.user.toString() !== req.user.id) {
+            return res.status(401).json({msg: 'User not authorized'});
+        }
 
         await rogal.deleteOne();
